% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sched-package}
\alias{sched}
\alias{sched-package}
\title{sched: Request Scheduler}
\description{
Offers classes and functions to contact web servers while enforcing scheduling rules required by the sites. The URL class makes it easy to construct a URL by providing parameters as a vector. The Request class allows to describes SOAP (Simple Object Access Protocol) or standard requests: URL, method (POST or GET), header, body. The Scheduler class controls the request frequency for each server address by mean of rules (Rule class). The RequestResult class permits to get the request status to handle error cases and the content.
}
\details{
\emph{sched} package.

\emph{sched} offers classes and functions to contact web servers while enforcing
scheduling rules required by the sites. The URL class makes it easy to
construct a URL by providing parameters as a vector. The Request class allows
to describes SOAP or standard requests: URL, method (POST or GET), header,
body. The Scheduler class controls the request frequency for each server
address by mean of rules (Rule class). The RequestResult class permits to get
the request status to handle error cases and the content.
}
\seealso{
\link{Rule}, \link{URL}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr} (\href{https://orcid.org/0000-0001-8177-4873}{ORCID})

}
